requirements/local.txt
appnope==0.1.3 ; python_version >= "3.10" and python_version < "3.11" and sys_platform == "darwin"
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "3.11"
asttokens==2.2.1 ; python_version >= "3.10" and python_version < "3.11"
attrs==22.2.0 ; python_version >= "3.10" and python_version < "3.11"
backcall==0.2.0 ; python_version >= "3.10" and python_version < "3.11"
black==22.12.0 ; python_version >= "3.10" and python_version < "3.11"
boto3-stubs==1.26.74 ; python_version >= "3.10" and python_version < "3.11"
boto3==1.26.74 ; python_version >= "3.10" and python_version < "3.11"
botocore-stubs==1.29.74 ; python_version >= "3.10" and python_version < "3.11"
botocore==1.29.74 ; python_version >= "3.10" and python_version < "3.11"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "3.11"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "3.11"
cfgv==3.3.1 ; python_version >= "3.10" and python_version < "3.11"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "3.11"
click==8.1.3 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and sys_platform == "win32" or python_version >= "3.10" and python_version < "3.11" and platform_system == "Windows"
coverage==7.1.0 ; python_version >= "3.10" and python_version < "3.11"
cryptography==39.0.1 ; python_version >= "3.10" and python_version < "3.11"
debugpy==1.6.6 ; python_version >= "3.10" and python_version < "3.11"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "3.11"
distlib==0.3.6 ; python_version >= "3.10" and python_version < "3.11"
django-cors-headers==3.13.0 ; python_version >= "3.10" and python_version < "3.11"
django-environ==0.9.0 ; python_version >= "3.10" and python_version < "3.11"
django-filter==22.1 ; python_version >= "3.10" and python_version < "3.11"
django-rest-knox==4.2.0 ; python_version >= "3.10" and python_version < "3.11"
django-storages==1.13.2 ; python_version >= "3.10" and python_version < "3.11"
django-stubs-ext==0.7.0 ; python_version >= "3.10" and python_version < "3.11"
django-stubs==1.14.0 ; python_version >= "3.10" and python_version < "3.11"
django==4.1.7 ; python_version >= "3.10" and python_version < "3.11"
djangorestframework-stubs==1.8.0 ; python_version >= "3.10" and python_version < "3.11"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "3.11"
dnspython==2.3.0 ; python_version >= "3.10" and python_version < "3.11"
exceptiongroup==1.1.0 ; python_version >= "3.10" and python_version < "3.11"
executing==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
factory-boy==3.2.1 ; python_version >= "3.10" and python_version < "3.11"
faker==15.3.4 ; python_version >= "3.10" and python_version < "3.11"
filelock==3.9.0 ; python_version >= "3.10" and python_version < "3.11"
flake8==6.0.0 ; python_version >= "3.10" and python_version < "3.11"
identify==2.5.18 ; python_version >= "3.10" and python_version < "3.11"
idna==3.4 ; python_version >= "3.10" and python_version < "3.11"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "3.11"
ipdb==0.13.11 ; python_version >= "3.10" and python_version < "3.11"
ipython==8.10.0 ; python_version >= "3.10" and python_version < "3.11"
isort==5.12.0 ; python_version >= "3.10" and python_version < "3.11"
jedi==0.18.2 ; python_version >= "3.10" and python_version < "3.11"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "3.11"
matplotlib-inline==0.1.6 ; python_version >= "3.10" and python_version < "3.11"
mccabe==0.7.0 ; python_version >= "3.10" and python_version < "3.11"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
mypy==0.991 ; python_version >= "3.10" and python_version < "3.11"
nodeenv==1.7.0 ; python_version >= "3.10" and python_version < "3.11"
packaging==23.0 ; python_version >= "3.10" and python_version < "3.11"
parso==0.8.3 ; python_version >= "3.10" and python_version < "3.11"
pathspec==0.11.0 ; python_version >= "3.10" and python_version < "3.11"
pexpect==4.8.0 ; python_version >= "3.10" and python_version < "3.11" and sys_platform != "win32"
pickleshare==0.7.5 ; python_version >= "3.10" and python_version < "3.11"
platformdirs==3.0.0 ; python_version >= "3.10" and python_version < "3.11"
pluggy==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
pre-commit==2.21.0 ; python_version >= "3.10" and python_version < "3.11"
prompt-toolkit==3.0.36 ; python_version >= "3.10" and python_version < "3.11"
psycopg2-binary==2.9.5 ; python_version >= "3.10" and python_version < "3.11"
ptyprocess==0.7.0 ; python_version >= "3.10" and python_version < "3.11" and sys_platform != "win32"
pure-eval==0.2.2 ; python_version >= "3.10" and python_version < "3.11"
pycodestyle==2.10.0 ; python_version >= "3.10" and python_version < "3.11"
pycparser==2.21 ; python_version >= "3.10" and python_version < "3.11"
pyflakes==3.0.1 ; python_version >= "3.10" and python_version < "3.11"
pygments==2.14.0 ; python_version >= "3.10" and python_version < "3.11"
pytest-django==4.5.2 ; python_version >= "3.10" and python_version < "3.11"
pytest==7.2.1 ; python_version >= "3.10" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.11"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "3.11"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "3.11"
requests==2.28.2 ; python_version >= "3.10" and python_version < "3.11"
s3transfer==0.6.0 ; python_version >= "3.10" and python_version < "3.11"
setuptools==67.3.2 ; python_version >= "3.10" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
sqlparse==0.4.3 ; python_version >= "3.10" and python_version < "3.11"
stack-data==0.6.2 ; python_version >= "3.10" and python_version < "3.11"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
traitlets==5.9.0 ; python_version >= "3.10" and python_version < "3.11"
types-awscrt==0.16.10 ; python_version >= "3.10" and python_version < "3.11"
types-pytz==2022.7.1.0 ; python_version >= "3.10" and python_version < "3.11"
types-pyyaml==6.0.12.6 ; python_version >= "3.10" and python_version < "3.11"
types-requests==2.28.11.13 ; python_version >= "3.10" and python_version < "3.11"
types-s3transfer==0.6.0.post5 ; python_version >= "3.10" and python_version < "3.11"
types-urllib3==1.26.25.6 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2022.7 ; python_version >= "3.10" and python_version < "3.11" and sys_platform == "win32"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "3.11"
virtualenv==20.19.0 ; python_version >= "3.10" and python_version < "3.11"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "3.11"
whitenoise==6.3.0 ; python_version >= "3.10" and python_version < "3.11"
